+++ src/Fantomas.Core/SyntaxOak.fs
@@ -1316,7 +1316,7 @@ type ElseIfNode(mElse: range, mIf: range, condition: Node, range) as elseIfNode 
                 (elseIfNode :> Node).AddBefore triviaNode
 
             member _.AddAfter(triviaNode: TriviaNode) =
-                    (elseIfNode :> Node).AddAfter triviaNode
+                (elseIfNode :> Node).AddAfter triviaNode
 
             member _.Children = Array.empty }