+++ src/Fantomas.Core/SyntaxOak.fs
@@ -1330,7 +1330,8 @@ type ElseIfNode(mElse: range, mIf: range, condition: Node, range) as elseIfNode 
 
             member _.AddBefore(triviaNode: TriviaNode) =
                 match triviaNode.Content with
-                | CommentOnSingleLine _ -> condition.AddBefore triviaNode
+                | CommentOnSingleLine _
+                | Newline -> condition.AddBefore triviaNode
                 | _ -> (elseIfNode :> Node).AddAfter triviaNode
 
             member _.AddAfter(triviaNode: TriviaNode) =
@@ -1348,7 +1349,6 @@ type ElseIfNode(mElse: range, mIf: range, condition: Node, range) as elseIfNode 
 
         member _.AddAfter(triviaNode: TriviaNode) =
             match triviaNode.Content with
-            | TriviaContent.Newline -> condition.AddBefore triviaNode
             | TriviaContent.LineCommentAfterSourceCode comment when lastNodeAfterIsLineCommentAfterSource ->
                 // If we already have a line comment after the `else if`, we cannot add another one.
                 // The next best thing would be to add it on the next line as content before of the condition.