+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 
-### Changed
+### Fixed
 * Duplicate newline in else if. [#2752](https://github.com/fsprojects/fantomas/issues/2752)
 
 ## [5.2.2] - 2023-02-18